New MQTT feature
Adding the possibility to deliver to an MQTT server instead of using
redis. The topics can be configured via ini file.
Assuming the MQTT server is running without any configuration, the
setup can be verified using (for mosquitto):

 # mosquitto_sub -v -h localhost -t raumklima/\*

Send a test publication using check_mqtt.py , which sends a demo value
to the configured server & topic.